corrige parametro do buscarporid

+++ src/main/java/com/reis/financeiro/service/RegistroService.java
@@ -46,7 +46,7 @@ import java.util.List;
             repository.delete(registro);
 
         }
-        public Registro buscarPorId(int id){
+        public Registro buscarPorId(long id){
             Registro registro = repository.findById(id).orElseThrow(()-> new RegistroNotFoundException());
             return registro;
         }
